added browserstack test, tweaking test options
