+++ karma.conf.ci.js
@@ -1,11 +1,10 @@
 /* eslint-disable camelcase */
 module.exports = function (config) {
     config.set({
-        autoWatch: false,
         singleRun: true
     });
 
-    if (process.env.USE_CLOUD === "true" && process.env.SAUCE_USERNAME) {
+    if (process.env.USE_CLOUD === "true") {
         require("./karma.conf.cloud.js")(config);
     } else {
         console.log("running default test on Chrome");
